Clean up legacy capability code and remove obsolete tests
- move non-active capability/model modules under legacy/src and rewire imports
- drop legacy tests and old capability test fixtures from tests/src
- consolidate embedding helpers into src/utils/embedding_utils and delete src/generate_embeddings.py
- remove unused visualization utils from active src
- vendor wikipedia Model into wikipedia/model.py to remove legacy dependency
- update cleanup decision docs and legacy README
