+++ legacy/pre_schema_pipeline/src/run_embedding_eval.py
@@ -10,21 +10,21 @@ from omegaconf import DictConfig
 from scipy import stats
 from tqdm import tqdm
 
-from legacy.utils.legacy_data_utils import get_run_id
-from src.lbo import calculate_lbo_error, fit_lbo
-from src.utils import constants
-from src.utils.capability_discovery_utils import (
+from legacy.src.lbo import calculate_lbo_error, fit_lbo
+from legacy.src.utils.capability_discovery_utils import (
     select_complete_capabilities,
 )
-from src.utils.capability_management_utils import (
+from legacy.src.utils.lbo_utils import get_lbo_train_set
+from legacy.utils.legacy_capability_management_utils import (
     get_previous_capabilities,
 )
-from src.utils.data_utils import check_cfg
-from src.utils.embedding_utils import (
+from legacy.utils.legacy_data_utils import get_run_id
+from legacy.utils.legacy_embedding_utils import (
     apply_dimensionality_reduction,
     generate_and_set_capabilities_embeddings,
 )
-from src.utils.lbo_utils import get_lbo_train_set
+from src.utils import constants
+from src.utils.data_utils import check_cfg
 
 
 @hydra.main(version_base=None, config_path="cfg", config_name="run_cfg")
