+++ legacy/pre_schema_pipeline/src/run_evaluation.py
@@ -6,15 +6,15 @@ import os
 import hydra
 from omegaconf import DictConfig
 
-from legacy.utils.legacy_data_utils import get_run_id
-from src.model import Model
-from src.utils import constants
-from src.utils.capability_discovery_utils import (
+from legacy.src.model import Model
+from legacy.src.utils.capability_discovery_utils import (
     select_complete_capabilities,
 )
-from src.utils.capability_management_utils import (
+from legacy.utils.legacy_capability_management_utils import (
     get_previous_capabilities,
 )
+from legacy.utils.legacy_data_utils import get_run_id
+from src.utils import constants
 from src.utils.data_utils import check_cfg
