+++ legacy/src/__init__.py
@@ -1,0 +1,1 @@
+"""Legacy core modules."""
