+++ legacy/src/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Legacy utility modules."""
