+++ legacy/src/utils/capability_utils.py
