+++ legacy/src/utils/templates.py
