--- tests/src/capabilities_t2/math/competition_math_combinatorial_challenges/_state.json
@@ -1,3 +1,0 @@
-{
-    "state": "initialized"
-}
