--- tests/src/resources/manual_capabilities.json
@@ -1,32 +1,0 @@
-{
-    "Algebra": {
-        "Linear and quadratic equations": "Focuses on equations involving first- and second-degree polynomials, including methods for solving them.",
-        "Algebraic expressions": "Involves simplifying and manipulating expressions using variables, constants, and operations.",
-        "Polynomials and factoring": "Covers identifying, expanding, and factoring polynomial expressions.",
-        "Systems of equations": "Studies techniques for solving multiple equations with multiple unknowns, such as substitution and elimination."
-    },
-    "Calculus and Analysis": {
-        "Differentiation": "Analyzes rates of change and slopes of functions using derivatives.",
-        "Integration": "Studies accumulation and area under curves through definite and indefinite integrals.",
-        "Sequences and series": "Investigates ordered lists of numbers and the behavior of their infinite sums.",
-        "Multivariable calculus": "Extends differentiation and integration to functions of several variables."
-    },
-    "Geometry and Trigonometry": {
-        "Euclidean geometry": "Studies points, lines, angles, and shapes in two- and three-dimensional space using classical axioms.",
-        "Coordinate geometry": "Connects algebra and geometry through graphs and equations on the Cartesian plane.",
-        "Trigonometric functions and identities": "Explores sine, cosine, tangent, and their relationships, often in relation to right triangles and circles.",
-        "Geometric transformations": "Examines operations like rotations, translations, reflections, and dilations of shapes."
-    },
-    "Discrete Mathematics and Logic": {
-        "Set theory and logic": "Covers foundational tools for reasoning, including sets, truth tables, and logical operators.",
-        "Combinatorics": "Studies counting methods, permutations, and combinations for finite structures.",
-        "Graph theory": "Explores structures made of nodes and edges, used in modeling networks and relationships.",
-        "Proof techniques": "Involves formal methods such as induction, contradiction, and direct proof to establish mathematical truths."
-    },
-    "Probability and Statistics": {
-        "Probability theory": "Studies outcomes and events in random experiments, using probability rules and models.",
-        "Combinatorial probability": "Applies counting techniques to calculate probabilities of complex events.",
-        "Random variables and distributions": "Investigates functions that assign numerical values to outcomes and their probability patterns.",
-        "Descriptive and inferential statistics": "Describes data sets and makes predictions or decisions based on sample data."
-    }
-}
