--- tests/src/seed_capabilities/math/competition_math_algebra/_state.json
@@ -1,3 +1,0 @@
-{
-    "state": "initialized"
-}
