--- tests/src/seed_capabilities/math/competition_math_algebra/capability.json
@@ -1,9 +1,0 @@
-{
-    "capability_name": "competition_math_algebra",
-    "capability_description": "The competition_math_algebra capability consists of 2931 challenging competition mathematics problems in algebra. Each problem has a full step-by-step solution which can be used to teach models to generate answer derivations and explanations. It has 5 levels.\n",
-    "source_dataset": "mathematics",
-    "capability_domain": "math",
-    "capability_instructions": "f\"\"\"Solve the following algebra math problem step by step. The last line of your response should be of the form \"ANSWER: $ANSWER\" (without quotes) where $ANSWER is the answer to the problem.\\n\\nProblem: {t[\"problem\"]}\\n\\nRemember to put your answer on its own line at the end in the form \"ANSWER:$ANSWER\" (without quotes) where $ANSWER is the answer to the problem, and you do not need to use a \\\\boxed command.\"\"\"",
-    "capability_data": [],
-    "capability_subject": "algebra"
-}
