+++ wikipedia/model.py
