+++ wikipedia/wikipedia_scraper.py
@@ -31,7 +31,7 @@ try:
     import sys
 
     sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
-    from src.model import Model
+    from wikipedia.model import Model
     from wikipedia.prompts import (
         SYSTEM_PROMPT_CAPABILITY_EVALUATION,
         SYSTEM_PROMPT_CATEGORIZATION,
